requirements-dev.txt
# ACOLYTE Development Requirements
# Install with: pip install -r requirements.txt -r requirements-dev.txt

pytest>=8.4.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
mypy>=1.8.0
black>=22.0.0
ruff>=0.12.0
types-pyyaml>=6.0.12
types-tree-sitter>=0.20.1
types-tree-sitter-languages>=1.10.0
types-aiofiles>=24.1.0
types-psutil>=7.0.0
types-requests>=2.32.4
weaviate-client>=3.26.7,<4.0.0
pytest-timeout>=2.4.0
